+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu Document Manager
 ===================================
 
 * dev-master
+    * HOTFIX      #118 Set path and node-name after renaming node
     * HOTFIX      #117 Execute rename in flush-event to avoid ItemNotFoundException
 
 * 0.10.0 (2017-06-28)