+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu Document Manager
 ===================================
 
 * dev-master
+    * HOTFIX      #119 Refactored version-subscriber to use uuid instead of paths
     * HOTFIX      #118 Set path and node-name after renaming node
     * HOTFIX      #117 Execute rename in flush-event to avoid ItemNotFoundException